refactor(find-wallet): use default modal variant, align close button, fold detail content into intercepted route
