+++ app/[locale]/wallets/find-wallet/_components/WalletDetailModal.tsx
@@ -1,38 +1,68 @@
 "use client"
 
+import { X } from "lucide-react"
+import type { StaticImageData } from "next/image"
 import { useRouter } from "next/navigation"
+import { useTranslations } from "next-intl"
 import type { ReactNode } from "react"
 
-import Modal from "@/components/ui/dialog-modal"
+import { Image } from "@/components/Image"
+import {
+  Dialog,
+  DialogClose,
+  DialogContent,
+  DialogTitle,
+} from "@/components/ui/dialog-modal"
 
 /**
  * Client shell for the intercepted wallet-detail route. The detail content is
  * server-rendered and passed as `children`; closing the modal pops the
  * intercepted route off the history stack, returning to the catalog.
+ *
+ * Composes the dialog primitives directly (rather than the `Modal` wrapper) so
+ * the wallet icon, name, and close button share one flex row — keeping the
+ * close button aligned with the visible title.
  */
 const WalletDetailModal = ({
   title,
+  image,
   children,
 }: {
   title: string
+  image: StaticImageData
   children: ReactNode
 }) => {
   const router = useRouter()
+  const t = useTranslations("common")
 
   return (
-    <Modal
+    <Dialog
       open
       size="lg"
-      variant="media"
       onOpenChange={(open) => {
         if (!open) router.back()
       }}
-      // Visible name lives in the detail content; this gives the dialog its
-      // accessible name without duplicating the heading.
-      title={<span className="sr-only">{title}</span>}
     >
-      {children}
-    </Modal>
+      <DialogContent>
+        <div className="flex items-center justify-between gap-4">
+          <div className="flex items-center gap-4">
+            <Image
+              src={image}
+              alt=""
+              className="size-14 shrink-0 object-contain"
+            />
+            <DialogTitle className="mt-0 text-h3">{title}</DialogTitle>
+          </div>
+          <DialogClose
+            aria-label={t("close")}
+            className="flex size-8 shrink-0 items-center justify-center rounded transition-colors hover:text-primary-hover"
+          >
+            <X size="20" />
+          </DialogClose>
+        </div>
+        {children}
+      </DialogContent>
+    </Dialog>
   )
 }
